Исправил: создал нестатический класс
