+++ src/org/skypro/skyshop/article/SearchEngine.java
@@ -1,5 +1,6 @@
 package org.skypro.skyshop.article;
 
+
 import java.util.*;
 
 public class SearchEngine {
@@ -14,17 +15,6 @@ public class SearchEngine {
     public void add(Searchable item) {
         searchableItems.add(item);
         count++;
-    }
-    public static class sortedTitleComparator implements Comparator<Searchable> {
-        @Override
-        public int compare(Searchable o1, Searchable o2) {
-            int lengthCompare = Integer.compare(o2.getName().length(), o1.getName().length());
-            if (lengthCompare != 0) {
-                return lengthCompare;
-            }
-            return o1.getName().compareTo(o2.getName());
-
-        }
     }
 
     public Set<Searchable> search(String term) {
